# swagger/swagger.yaml
basePath: /api/
consumes:
- application/json
definitions:
...
  OfTheDayModel:
    properties:
      date:
        description: The date when this author / quote should be 'of the day'
        example: "2020-06-12"
        type: string
        x-go-name: Date
      id:
        description: The id of the author / quote
        example: 1
        format: int64
        type: integer
        x-go-name: Id
      language:
        default: English
        description: The language of this author / quote
        example: icelandic
        type: string
        x-go-name: Language
    type: object
    x-go-name: ofTheDayModel
    x-go-package: github.com/Skjaldbaka17/quotes-api/swagger
  OrderConfiguration:
    properties:
      maximum:
        description: |-
          Where to end the ordering (if empty it ends at the logical end, for example end at the highest quoteid for quoteid ascending order).
          Note this key is always a string.
        example: "11"
        type: string
        x-go-name: Maximum
      minimum:
        description: |-
          Where to start the ordering (if empty it starts from beginning, for example start at 1 for quoteid ascending order).
          Note this key is always a string.
        example: "10"
        type: string
        x-go-name: Minimum
      orderBy:
        description: What to order by, 'quoteId', 'popularity' or 'length'
        example: popularity
        type: string
        x-go-name: OrderBy
      reverse:
        description: Whether to order the list in reverse or not (true is Descending
          and false is Ascending, false is default)
        example: true
        type: boolean
        x-go-name: Reverse
    type: object
    x-go-name: orderConfigListQuotesModel
    x-go-package: github.com/Skjaldbaka17/quotes-api/swagger
  QodViewAPIModel:
    properties:
      authorId:
        description: The author's id
        format: int64
        type: integer
        x-go-name: AuthorId
      date:
        description: The date when this quote was the quote of the day
        example: "2021-06-12T00:00:00Z"
        type: string
        x-go-name: Date
      isIcelandic:
        description: |-
          Whether the quote is in icelandic
          false
        type: boolean
        x-go-name: IsIcelandic
      name:
        description: Name of the author
        example: Muhammad Ali
        type: string
        x-go-name: Name
      quote:
        description: The quote for the day
        example: Float like a butterfly, sting like a bee
        type: string
        x-go-name: Quote
      quoteId:
        description: The quote's id
        example: 582676
        format: int64
        type: integer
        x-go-name: QuoteId
    type: object
    x-go-package: github.com/Skjaldbaka17/quotes-api/structs
  SearchViewAPIModel:
    properties:
      authorId:
        description: The author's id
        example: 24952
        format: int64
        type: integer
        uniqueItems: true
        x-go-name: AuthorId
      isIcelandic:
        description: Whether or not this quote is in Icelandic or not
        example: false
        type: boolean
        x-go-name: IsIcelandic
      name:
        description: Name of author
        example: Muhammad Ali
        type: string
        x-go-name: Name
      quote:
        description: The quote
        example: Float like a butterfly, sting like a bee.
        type: string
        x-go-name: Quote
      quoteId:
        description: The quote's id
        example: 582676
        format: int64
        type: integer
        uniqueItems: true
        x-go-name: QuoteId
    type: object
    x-go-package: github.com/Skjaldbaka17/quotes-api/structs
...
  TopicViewAPIModel:
    properties:
      authorId:
        description: The author's id
        example: 24952
        format: int64
        type: integer
        uniqueItems: true
        x-go-name: AuthorId
      isIcelandic:
        description: Whether or not this quote is in Icelandic or not
        example: false
        type: boolean
        x-go-name: IsIcelandic
      name:
        description: Name of author
        example: Muhammad Ali
        type: string
        x-go-name: Name
      quote:
        description: The quote
        example: Float like a butterfly, sting like a bee.
        type: string
        x-go-name: Quote
      quoteId:
        description: The quote's id
        example: 582676
        format: int64
        type: integer
        uniqueItems: true
        x-go-name: QuoteId
      topicId:
        description: The topic's id (if topic id / name not supplied this will return
          a zero id)
        example: 10
        format: int64
        type: integer
        x-go-name: TopicId
      topicName:
        description: The topic's name (if topic id / name not supplied this will return
          empty string "")
        example: inspirational
        type: string
        x-go-name: TopicName
    type: object
    x-go-package: github.com/Skjaldbaka17/quotes-api/structs
...
host: quotel-api.com
info:
  contact:
    email: skjaldbaka17@gmail.com
    name: Þórður Ágústsson
  description: |-
    Quotel API offers a complete feature rich REST API access to all your quotes needs. This is the documentation of our quotes API.
    To use it you first need to sign up here https://somethingsomethingdarkside.com to get an authentication api key.
  title: quotes-api.
  version: 1.0.0
paths:
...
  /quotes:
    post:
      description: Get quotes by their ids
      operationId: GetQuotes
      parameters:
      - description: |-
          The structure of the request to get quotes. There are two ways to use this route. 1. Send the ids of the quotes to be
          retrieved or 2. send the id of the author of the quotes you want (if you use this option the quotes are paginated)
        in: body
        name: Body
        required: true
        schema:
          properties:
            apiKey:
              description: The api-key you use to access the api
              example: 91fd6d19-2c32-4081-8729-4d9786d43b95
              type: string
              x-go-name: ApiKey
            authorId:
              description: The id of the author of the quotes you want.
              example: 24952
              format: int64
              type: integer
              x-go-name: AuthorId
            ids:
              description: The list of quotes's ids you want
              example:
              - 582676
              - 443976
              items:
                format: int64
                type: integer
              type: array
              x-go-name: Ids
            page:
              description: If using authorId the response is paged. This parameter
                controls the page you are asking for, starts with 0.
              example: 0
              format: int64
              minimum: 0
              type: integer
              x-go-name: Page
            pageSize:
              default: 25
              description: If using authorId the response is paged. This parameter
                controls the number of Authors to be returned on each "page"
              example: 30
              format: int64
              maximum: 200
              minimum: 1
              type: integer
              x-go-name: PageSize
          required:
          - apiKey
          type: object
      responses:
        "200":
          $ref: '#/responses/searchViewsResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - QUOTES
  /quotes/list:
    post:
      description: Get list of quotes according to some ordering / parameters
      operationId: GetQuotesList
      parameters:
      - description: The structure of the request for getting a list of quotes
        in: body
        name: Body
        schema:
          properties:
            apiKey:
              description: The api-key you use to access the api
              example: 91fd6d19-2c32-4081-8729-4d9786d43b95
              type: string
              x-go-name: ApiKey
            language:
              description: |-
                Only return quotes that have quotes in the given language ("english" or "icelandic") if left empty then no constraint
                is set on the quotes' language.
              example: English
              type: string
              x-go-name: Language
            orderConfig:
              $ref: '#/definitions/OrderConfiguration'
            page:
              description: Response is paged. This parameter controls the page you
                are asking for, starts with 0.
              example: 0
              format: int64
              minimum: 0
              type: integer
              x-go-name: Page
            pageSize:
              default: 25
              description: Response is paged. This parameter controls the number of
                Quotes to be returned on each "page"
              example: 30
              format: int64
              maximum: 200
              minimum: 1
              type: integer
              x-go-name: PageSize
          required:
          - apiKey
          type: object
      responses:
        "200":
          $ref: '#/responses/searchViewsResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - QUOTES
  /quotes/qod:
    post:
      description: gets the quote of the day
      operationId: GetQuoteOfTheDay
      parameters:
      - description: The structure of the request for getting the author / quote of
          the day
        in: body
        name: Body
        schema:
          properties:
            apiKey:
              description: The api-key you use to access the api
              example: 91fd6d19-2c32-4081-8729-4d9786d43b95
              type: string
              x-go-name: ApiKey
            language:
              default: English
              description: Get the author / quote of the day for the given language
                ("icelandic" or "english")
              example: English
              type: string
              x-go-name: Language
          required:
          - apiKey
          type: object
      responses:
        "200":
          $ref: '#/responses/qodResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - QUOTES
  /quotes/qod/history:
    post:
      description: Gets the history for the quotes of the day
      operationId: GetQODHistory
      parameters:
      - description: The structure of the request for getting the history of AODs
          / QODs
        in: body
        name: Body
        schema:
          items:
            properties:
              apiKey:
                description: The api-key you use to access the api
                example: 91fd6d19-2c32-4081-8729-4d9786d43b95
                type: string
                x-go-name: ApiKey
              language:
                default: English
                description: Get the history of the AODS / QODs for the given language
                  ("icelandic" or "english")
                example: icelandic
                type: string
                x-go-name: Language
              minimum:
                description: The earliest date to return. All authors / quotes between
                  minimum and today will be returned.
                example: "2020-12-21"
                type: string
                x-go-name: Minimum
            required:
            - apiKey
            type: object
          type: array
      responses:
        "200":
          $ref: '#/responses/qodHistoryResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - QUOTES
  /quotes/qod/new:
    post:
      description: Sets the quote of the day for the given dates
      operationId: SetQuoteOfTheDay
      parameters:
      - description: The structure of the request for setting the QOD
        in: body
        name: Body
        schema:
          properties:
            apiKey:
              description: The api-key you use to access the api
              example: 91fd6d19-2c32-4081-8729-4d9786d43b95
              type: string
              x-go-name: ApiKey
            qods:
              items:
                $ref: '#/definitions/OfTheDayModel'
              type: array
              x-go-name: Qods
          required:
          - apiKey
          type: object
      responses:
        "200":
          $ref: '#/responses/successResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - QUOTES
  /quotes/random:
    post:
      description: Get a random quote according to the given parameters
      operationId: GetRandomQuote
      parameters:
      - description: The structure of the request for a random quote
        in: body
        name: Body
        schema:
          properties:
            apiKey:
              description: The api-key you use to access the api
              example: 91fd6d19-2c32-4081-8729-4d9786d43b95
              type: string
              x-go-name: ApiKey
            authorid:
              description: The random quote returned must be from the author with
                the given authorId
              example: 24952
              format: int64
              type: integer
              x-go-name: Authorid
            language:
              description: The random quote returned must be in the given language
              example: English
              type: string
              x-go-name: Language
            searchString:
              description: The random quote returned must contain a match with the
                searchstring
              example: float
              type: string
              x-go-name: SearchString
            topicId:
              description: The random quote returned must be a part of the topic with
                the given topicId
              example: 10
              format: int64
              type: integer
              x-go-name: TopicId
          required:
          - apiKey
          type: object
      responses:
        "200":
          $ref: '#/responses/topicViewResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - QUOTES
...
produces:
- application/json
responses:
...
  incorrectBodyStructureResponse:
    description: Data structure representing the error response to a wrongly structured
      request body
    schema:
      properties:
        message:
          description: The error message
          example: request body is not structured correctly.
          type: string
          x-go-name: Message
      type: object
...
  internalServerErrorResponse:
    description: Data structure representing the error response to an internal server
      error
    schema:
      properties:
        message:
          description: The error message
          example: Please try again later.
          type: string
          x-go-name: Message
      type: object
...
  qodHistoryResponse:
    description: Data structure representing the response for the history of QODS
    schema:
      items:
        $ref: '#/definitions/QodViewAPIModel'
      type: array
  qodResponse:
    description: Data structure representing the response for the quote of the day
    schema:
      $ref: '#/definitions/QodViewAPIModel'
...
  searchViewsResponse:
    description: Data structure representing the response for quotes
    schema:
      items:
        $ref: '#/definitions/SearchViewAPIModel'
      type: array
  successResponse:
    description: ""
    schema:
      properties:
        message:
          example: This request was a success
          type: string
          x-go-name: Message
        statusCode:
          description: HTTP status code
          example: 200
          format: int64
          type: integer
          x-go-name: StatusCode
      type: object
  topicViewResponse:
    description: Data structure representing the response for a quote based on a particular
      topic
    schema:
      $ref: '#/definitions/TopicViewAPIModel'
...
schemes:
- http
securityDefinitions:
  apikey:
    description: To use the api you need an API key. The basic API key lets you make
      up to 100 requests an hour.
    in: body
    name: apiKey
    type: apiKey
swagger: "2.0"
...